Added volume support
All data and code is now pulled in through a volume. As a consequence of this, building the app can no longer be done directly from the Dockerfile, so has been moved into its own `init` command in the entry point.

While I was at it, I have also created a command for starting sneakers to the entrypoint.