Basic Dockerfile for node-based run.
Since this is such a simple tool, we simply copy all the code into the container and then run npm install. This isn't ideal, we should really be using a volume to copy in the code, but it turns out that this causes problems with storing the output of npm install. For now, we will live with this.