Fix locking issue in dev mode
The SQLite adapter by default opens a read lock only when running a new job.
As our submission jobs now write to the database, too, we need to take out a write lock right-away to avoid transaction rollback on concurrent writes. Not ideal, as it effectively sequentialises jobs.

For the moment, I've only added this to the dev environment's SQLite3 usage. I may also have to do this for production Postgres usage. I would really like to avoid this as much as possible and find an alternative solution instead.